no user can delete comment by another user

+++ portfolio/src/main/java/com/google/sps/servlets/deleteComment.java
@@ -1,6 +1,8 @@
 package com.google.sps.servlets;
 
 import com.google.appengine.api.datastore.*;
+import com.google.appengine.api.users.UserService;
+import com.google.appengine.api.users.UserServiceFactory;
 import com.google.gson.Gson;
 import com.google.sps.data.comment;
 import java.io.IOException;
@@ -17,7 +19,16 @@ public final class deleteComment extends HttpServlet {
   @Override
   public void doPost(HttpServletRequest request, HttpServletResponse response)
     throws IOException {
+    UserService userService = UserServiceFactory.getUserService();
+
+    // if user is not loggedin the comment can't be deleted
+    if (!userService.isUserLoggedIn()) {
+      return;
+    }
     long id = Long.parseLong(request.getParameter("id"));
+
+    // a user cannot delete comment by another user
+    if (id != Long.parseLong(userService.getCurrentUser().getUserId())) return;
     Key commentEntityKey = KeyFactory.createKey("comment", id);
     DatastoreService datastore = DatastoreServiceFactory.getDatastoreService();
     datastore.delete(commentEntityKey);
